requirements.txt
Django==1.5.4
Django-Select2==4.2.1
PIL==1.1.7
Paver==1.2.1
South==0.8.2
django-appconf==0.6
django-bootstrap-toolkit==2.15.0
django-compressor==1.3
django-forms-bootstrap==2.0.3.post1
django-geoexplorer==3.0.4
django-jsonfield==0.8.12
django-reversion==1.7.1
django-singleton==0.1.7
django-stronghold==0.2.2
django-tastypie==0.11.0
django-user-accounts==1.0b13
eventlog==0.6.2
gamification==0.0.0.1
jsonfield==0.9.17
metron==1.1
pinax-theme-bootstrap==3.0a9
psycopg2==2.5.1
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2012j
requests==1.2.3
six==1.4.1
wsgiref==0.1.2
